flask-app/requirements.txt
opencv-python<=3.5
opencv-contrib-python<=3.5
imageio==2.5.0
tqdm==4.31.1
numpy==1.16.3
matplotlib==3.0.3
scipy
Pillow
cython
scikit-image
tensorflow==1.3.0
keras==2.0.8
h5py
imgaug
IPython[all]

